Added border blending for smooth circles
I also fixed a minor off-by-one error that caused the border of the image
to be too small by one pixel horizontally and vertically. Blending made the
issue more obvious so it had to be fixed.

Finally, blending is done on the color value for r, g, and b post-rounding
but the val_to_rgb function could be modified to take the blend amount
so that the blending is applied before rounding occurs.